+++ src/components/ui/popover.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import * as PopoverPrimitive from '@radix-ui/react-popover';
 
 import { cn } from '@/src/lib/utils';
+import { PopoverClose } from '@radix-ui/react-popover';
 
 function Popover({
   ...props
@@ -43,4 +44,4 @@ function PopoverAnchor({
   return <PopoverPrimitive.Anchor data-slot='popover-anchor' {...props} />;
 }
 
-export { Popover, PopoverTrigger, PopoverContent, PopoverAnchor };
+export { Popover, PopoverTrigger, PopoverContent, PopoverAnchor, PopoverClose };