+++ src/flows/ContractorOnboarding/components/OnboardingSubmit.tsx
@@ -12,18 +12,11 @@ export function OnboardingSubmit({
   const { formId } = useContractorOnboardingContext();
   const { components } = useFormFields();
 
-  const CustomButton = components?.button;
-  if (CustomButton) {
-    return (
-      <CustomButton {...props} form={formId}>
-        {children}
-      </CustomButton>
-    );
-  }
+  const CustomButton = components?.button || Button;
 
   return (
-    <Button {...props} form={formId}>
+    <CustomButton {...props} form={formId}>
       {children}
-    </Button>
+    </CustomButton>
   );
 }